pull JsonLfDecoders out from JsonLfReader.Decoders
